Add filter modal for small screens
On small screens, the shopping page now has a filter button which opens
a modal with the filtering options found in the sidebar on
larger screens.  The filter is refactored into a reusable component.
Finally, the custom price filter is now a controlled component that
reacts instantaneously, as opposed to a form that required clicking on
a submit button.